[DsBrowserPage] moved image viewer to separated module
