+++ metaspace/browser/setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='sm',
+    version='0.0.1',
+    description='High throughput molecules annotation for imaging mass spectrometry data sets',
+    url='https://github.com/metaspace2020/metaspace/tree/master/metaspace/engine',
+    author='Alexandrov Team, EMBL',
+    author_email='vitaly.kovalev@embl.de',
+    packages=find_packages(),
+    install_requires=[],
+)
