+++ metaspace/webapp/src/modules/Datasets/imzml/DatasetBrowserPage.scss
@@ -9,22 +9,6 @@ $header-height: 64px;
     width: 70px;
   }
 
-  .range-slider{
-    width: 200px;
-  }
-
-  .ds-intensities-wrapper{
-    @apply flex justify-between;
-
-    .icon-check{
-      z-index: 1;
-    }
-  }
-
-  .ds-viewer-controls-wrapper{
-    @apply absolute top-0 right-0 py-3 mr-2 h-full box-border flex flex-col justify-between items-end w-0;
-  }
-
   .ion-image-holder{
     @apply flex flex-col items-center justify-start relative;
     min-height: 500px;
@@ -84,10 +68,6 @@ $header-height: 64px;
 
   h1 {
     @apply text-4xl leading-12 font-display;
-  }
-
-  .dataset-comparison-grid-item-header{
-    @apply flex justify-end;
   }
 
   .dataset-browser-wrapper{
