+++ metaspace/webapp/src/modules/Datasets/imzml/DatasetBrowserPage.tsx
@@ -5,20 +5,10 @@ import { getDatasetByIdQuery, GetDatasetByIdQuery } from '../../../api/dataset'
 import { annotationListQuery } from '../../../api/annotation'
 import config from '../../../lib/config'
 import safeJsonParse from '../../../lib/safeJsonParse'
-import MainImage from '../../Annotations/annotation-widgets/default/MainImage.vue'
-import MainImageHeader from '../../Annotations/annotation-widgets/default/MainImageHeader.vue'
-import FadeTransition from '../../../components/FadeTransition'
-import RangeSlider from '../../../components/Slider/RangeSlider.vue'
-import IonIntensity from '../../ImageViewer/IonIntensity.vue'
-import { loadPngFromUrl, processIonImage, renderScaleBar } from '../../../lib/ionImageRendering'
-import { get } from 'lodash-es'
-import createColormap from '../../../lib/createColormap'
-import getColorScale from '../../../lib/getColorScale'
-import { THUMB_WIDTH } from '../../../components/Slider'
-import { periodicTable } from './periodicTable'
-import Vue from 'vue'
 import DatasetBrowserSpectrumChart from './DatasetBrowserSpectrumChart'
 import './DatasetBrowserPage.scss'
+import SimpleIonImageViewer from './SimpleIonImageViewer'
+import { calculateMzFromFormula } from '../../../lib/formulaParser'
 
 interface DatasetBrowserProps {
   className: string
@@ -32,14 +22,10 @@ interface DatasetBrowserState {
   mzmScoreFilter: number | undefined
   mzmPolarityFilter: number | undefined
   mzmScaleFilter: string | undefined
-  scaleIntensity: boolean
   ionImageUrl: any
   sampleData: any[]
   chartLoading: boolean
   imageLoading: boolean
-  ionImage: any
-  rangeSliderStyle: any
-  imageSettings: any
   metadata: any
   annotation: any
   x: number | undefined
@@ -69,13 +55,9 @@ export default defineComponent<DatasetBrowserProps>({
       mzmPolarityFilter: undefined,
       mzmScaleFilter: undefined,
       metadata: undefined,
-      ionImage: undefined,
       annotation: undefined,
-      rangeSliderStyle: undefined,
       chartLoading: false,
       imageLoading: false,
-      scaleIntensity: false,
-      imageSettings: undefined,
       moleculeFilter: undefined,
       x: undefined,
       y: undefined,
@@ -168,226 +150,6 @@ export default defineComponent<DatasetBrowserProps>({
       }
     }
 
-    const handleIonIntensityChange = async(intensity: number, type: string) => {
-      let minScale = state.imageSettings.userScaling[0]
-      let maxScale = state.imageSettings.userScaling[1]
-
-      if (type === 'min') {
-        minScale = intensity / state.imageSettings.intensity.max.image
-        minScale = minScale > 1 ? 1 : minScale
-        minScale = minScale > maxScale ? maxScale : minScale
-        minScale = minScale < 0 ? 0 : minScale
-      } else {
-        maxScale = intensity / state.imageSettings.intensity.max.image
-        maxScale = maxScale > 1 ? 1 : maxScale
-        maxScale = maxScale < 0 ? 0 : maxScale
-        maxScale = maxScale < minScale ? minScale : maxScale
-      }
-
-      handleUserScalingChange([minScale, maxScale])
-    }
-
-    const handleUserScalingChange = async(userScaling: any) => {
-      state.imageSettings.userScaling = userScaling
-
-      const minScale =
-        state.imageSettings.intensity?.min?.status === 'LOCKED'
-          ? userScaling[0] * (1
-          - (state.imageSettings.intensity.min.user / state.imageSettings.intensity.max.image))
-          + (state.imageSettings.intensity.min.user / state.imageSettings.intensity.max.image)
-          : userScaling[0]
-
-      const maxScale = userScaling[1] * (state.imageSettings.intensity?.max?.status === 'LOCKED'
-        ? state.imageSettings.intensity.max.user / state.imageSettings.intensity.max.image : 1)
-      const scale = [minScale, maxScale]
-      state.imageSettings.imageScaledScaling = scale
-
-      Vue.set(state.imageSettings, 'intensity', {
-        ...state.imageSettings.intensity,
-        min:
-          {
-            ...state.imageSettings.intensity.min,
-            scaled: state.imageSettings.intensity.max.image * userScaling[0],
-          },
-        max:
-          {
-            ...state.imageSettings.intensity.max,
-            scaled: state.imageSettings.intensity.max.image * userScaling[1],
-          },
-      })
-    }
-
-    const handleIonIntensityLockChange = async(value: number, type: string) => {
-      const minLocked = type === 'min' ? value : state.imageSettings.lockedIntensities[0]
-      const maxLocked = type === 'max' ? value : state.imageSettings.lockedIntensities[1]
-      const lockedIntensities = [minLocked, maxLocked]
-      const intensity = getIntensity(state.ionImage,
-        lockedIntensities)
-
-      if (intensity && intensity.max && maxLocked && intensity.max.status === 'LOCKED') {
-        intensity.max.scaled = maxLocked
-        intensity.max.user = maxLocked
-        intensity.max.clipped = maxLocked
-      }
-
-      if (intensity && intensity.min && minLocked && intensity.min.status === 'LOCKED') {
-        intensity.min.scaled = minLocked
-        intensity.min.user = minLocked
-        intensity.min.clipped = minLocked
-      }
-
-      if (intensity && intensity.min && intensity.min.status !== 'LOCKED'
-      ) {
-        state.imageSettings.imageScaledScaling = [0, state.imageSettings.imageScaledScaling[1]]
-      }
-      if (intensity && intensity.max && intensity.max.status !== 'LOCKED') {
-        state.imageSettings.imageScaledScaling = [state.imageSettings.imageScaledScaling[0], 1]
-      }
-
-      state.imageSettings.lockedIntensities = lockedIntensities
-      state.imageSettings.intensity = intensity
-      state.imageSettings.userScaling = [0, 1]
-    }
-
-    const buildRangeSliderStyle = (scaleRange: number[] = [0, 1]) => {
-      const width = 190
-      const activeColorMap = state.imageSettings.colormap
-      const ionImage = state.ionImage
-      const cmap = createColormap(activeColorMap)
-      const { range } = getColorScale(activeColorMap)
-      const { scaledMinIntensity, scaledMaxIntensity } = ionImage || {}
-      const minColor = range[0]
-      const maxColor = range[range.length - 1]
-      const gradient = scaledMinIntensity === scaledMaxIntensity
-        ? `linear-gradient(to right, ${range.join(',')})`
-        : ionImage ? `url(${renderScaleBar(ionImage, cmap, true)})` : ''
-
-      state.imageSettings.colorBar = {
-        minColor,
-        maxColor,
-        gradient,
-      }
-
-      const [minScale, maxScale] = scaleRange
-      const minStop = Math.ceil(THUMB_WIDTH + ((width - THUMB_WIDTH * 2) * minScale))
-      const maxStop = Math.ceil(THUMB_WIDTH + ((width - THUMB_WIDTH * 2) * maxScale))
-      state.rangeSliderStyle = {
-        background: [
-          `0px / ${minStop}px 100% linear-gradient(${minColor},${minColor}) no-repeat`,
-          `${minStop}px / ${maxStop - minStop}px 100% ${gradient} repeat-y`,
-          `${minColor} ${maxStop}px / ${width - maxStop}px 100% linear-gradient(${maxColor},${maxColor}) no-repeat`,
-        ].join(','),
-      }
-    }
-
-    const setIonImage = async() => {
-      if (state.ionImageUrl) {
-        const ionImagePng = await loadPngFromUrl(state.ionImageUrl)
-        const isotopeImage = get(state.annotation, 'isotopeImages[0]')
-        const { minIntensity, maxIntensity } = isotopeImage
-        state.ionImage = await processIonImage(ionImagePng, minIntensity, maxIntensity,
-          state.imageSettings.scaleType)
-        state.imageSettings.intensity = getIntensity(state.ionImage)
-        buildRangeSliderStyle()
-      } else {
-        state.ionImage = null
-      }
-    }
-
-    const parseFormula = (formula: string) => {
-      const regexp = /(?<element>[A-Z][a-z]{0,2})(?<n>[0-9]*)/ig
-      const elements : any = {}
-      Array.from(formula.matchAll(regexp), (res, idx) => {
-        if (res.groups && res.groups.element && res.groups.element.length > 1) {
-          if (periodicTable[res.groups.element] === undefined) {
-            const auxElement = res.groups.element.substring(1, 2)
-            const auxElementCount = parseFormula(auxElement
-              + (res.groups && res.groups.n ? parseInt(res.groups.n, 10) : 1))
-
-            if (Object.keys(elements).includes(auxElement)) {
-              elements[auxElement] += auxElementCount[auxElement]
-            } else {
-              elements[auxElement] = auxElementCount[auxElement]
-            }
-
-            res.groups.element = res.groups.element.substring(0, 1)
-            res.groups.n = '1'
-          }
-        }
-
-        if (res.groups && res.groups.element in Object.keys(elements)) {
-          elements[res.groups.element] = elements[res.groups.element]
-            + (res.groups && res.groups.n ? parseInt(res.groups.n, 10) : 1)
-        } else if (res.groups && !(res.groups.element in Object.keys(elements))) {
-          elements[res.groups.element] = (res.groups && res.groups.n ? parseInt(res.groups.n, 10) : 1)
-        }
-      })
-
-      return elements
-    }
-
-    const formatFormula = (elements: any) => {
-      let formula = ''
-      Object.keys(elements).sort().forEach((elementKey: string) => {
-        const element = elements[elementKey]
-        if (element > 0) {
-          formula += elementKey + element
-        }
-      })
-      return formula
-    }
-
-    const calculateMzFromFormula = (molecularFormula: string) => {
-      const ionFormula = generateIonFormula(molecularFormula)
-      const ionElements = parseFormula(ionFormula)
-      let mz = 0
-
-      Object.keys(ionElements).forEach((elementKey: string) => {
-        const nOfElements = ionElements[elementKey]
-        if (periodicTable[elementKey]) {
-          const mass = periodicTable[elementKey][2][0]
-          mz += nOfElements * mass
-        }
-      })
-
-      return mz
-    }
-
-    const generateIonFormula = (molecularFormula: string) => {
-      const cleanFormula = molecularFormula.toUpperCase().trim().replace(/\s/g, '')
-      const regexpFormulas = /(?<formula>\w+)(?<adducts>([+-]\w+)*)/ig
-      const match = regexpFormulas.exec(cleanFormula)
-      const formula = match && match.groups ? match.groups.formula : ''
-      const adducts : string[] = []
-
-      if (match && match.groups && match.groups.adducts) {
-        const regexpAdduct = /([+-]\w+)/ig
-        Array.from(match.groups.adducts.matchAll(regexpAdduct),
-          (res, idx) => {
-            adducts.push(res[0])
-          })
-      }
-
-      const ionElements = parseFormula(formula)
-
-      adducts.forEach((adduct: string) => {
-        const elem = adduct.replace(/[^a-zA-Z]/g, '')
-        if (adduct.indexOf('+') !== -1) {
-          if (Object.keys(ionElements).includes(elem)) {
-            ionElements[elem] += 1
-          } else {
-            ionElements[elem] = 1
-          }
-        } else if (adduct.indexOf('-') !== -1) {
-          if (Object.keys(ionElements).includes(elem)) {
-            ionElements[elem] -= 1
-          }
-        }
-      })
-
-      return formatFormula(ionElements)
-    }
-
     const requestIonImage = async(mzValue : number | undefined = state.mzmScoreFilter) => {
       // @ts-ignore
       const inputPath: string = dataset.value.inputPath.replace('s3a:', 's3:')
@@ -436,16 +198,28 @@ export default defineComponent<DatasetBrowserProps>({
         state.mzmPolarityFilter = ppm
         state.mzmScaleFilter = 'ppm'
         await requestIonImage()
-        if (!state.imageSettings) {
-          startImageLoaderSettings()
-        }
-        await setIonImage()
         buildMetadata(dataset.value)
         if (state.x !== undefined && state.y !== undefined) {
           await requestSpectrum(state.x, state.y)
         }
       }
       queryOptions.enabled = false
+    })
+
+    const metadata : any = computed(() => {
+      let metadataAux = {}
+
+      if (dataset.value) {
+        metadataAux = {
+          Submitter: dataset.value.submitter,
+          PI: dataset.value.principalInvestigator,
+          Group: dataset.value.group,
+          Projects: dataset.value.projects,
+        }
+        metadataAux = Object.assign(safeJsonParse(dataset.value.metadataJson), metadataAux)
+      }
+
+      return metadataAux
     })
 
     const buildMetadata = (dataset: any) => {
@@ -459,19 +233,23 @@ export default defineComponent<DatasetBrowserProps>({
     }
 
     const getPixelSizeX = () => {
-      if (state.metadata && state.metadata.MS_Analysis != null
-        && state.metadata.MS_Analysis.Pixel_Size != null) {
-        return state.metadata.MS_Analysis.Pixel_Size.Xaxis
+      if (metadata.value && metadata.value.MS_Analysis != null
+        && metadata.value.MS_Analysis.Pixel_Size != null) {
+        return metadata.value.MS_Analysis.Pixel_Size.Xaxis
       }
       return 0
     }
 
     const getPixelSizeY = () => {
-      if (state.metadata && state.metadata.MS_Analysis != null
-        && state.metadata.MS_Analysis.Pixel_Size != null) {
-        return state.metadata.MS_Analysis.Pixel_Size.Yaxis
+      if (metadata.value && metadata.value.MS_Analysis != null
+        && metadata.value.MS_Analysis.Pixel_Size != null) {
+        return metadata.value.MS_Analysis.Pixel_Size.Yaxis
       }
       return 0
+    }
+
+    const handlePixelSelect = (coordinates: any) => {
+      requestSpectrum(coordinates.x, coordinates.y)
     }
 
     const renderBrowsingFilters = () => {
@@ -622,100 +400,6 @@ export default defineComponent<DatasetBrowserProps>({
       )
     }
 
-    const handlePixelSelect = (coordinates: any) => {
-      requestSpectrum(coordinates.x, coordinates.y)
-    }
-
-    const handleImageMove = ({ zoom, xOffset, yOffset }: any) => {
-      state.imageSettings.imagePosition.zoom = zoom
-      state.imageSettings.imagePosition.xOffset = xOffset
-      state.imageSettings.imagePosition.yOffset = yOffset
-    }
-
-    const handleColormapChange = (colormap: string) => {
-      state.imageSettings.colormap = colormap
-      buildRangeSliderStyle()
-    }
-
-    const handleScaleTypeChange = (scaleType: string) => {
-      state.imageSettings.scaleType = scaleType
-      buildRangeSliderStyle()
-    }
-
-    const handleScaleBarColorChange = (color: string) => {
-      state.imageSettings.scaleBarColor = color
-    }
-
-    const getIntensityData = (
-      image: number, clipped: number, scaled: number, user: number, quantile: number, isLocked?: boolean,
-    ) => {
-      const isClipped = quantile > 0 && quantile < 1 && user === image
-      return {
-        image,
-        clipped,
-        scaled,
-        user,
-        quantile,
-        status: isLocked ? 'LOCKED' : isClipped ? 'CLIPPED' : undefined,
-      }
-    }
-
-    const getIntensity = (ionImage: any, lockedIntensities: any = []) => {
-      if (ionImage !== null) {
-        const {
-          minIntensity, maxIntensity,
-          clippedMinIntensity, clippedMaxIntensity,
-          scaledMinIntensity, scaledMaxIntensity,
-          userMinIntensity, userMaxIntensity,
-          lowQuantile, highQuantile,
-        } = ionImage || {}
-        const [lockedMin, lockedMax] = lockedIntensities
-
-        return {
-          min: getIntensityData(
-            minIntensity,
-            clippedMinIntensity,
-            scaledMinIntensity,
-            userMinIntensity,
-            lowQuantile,
-            lockedMin !== undefined,
-          ),
-          max: getIntensityData(
-            maxIntensity,
-            clippedMaxIntensity,
-            scaledMaxIntensity,
-            userMaxIntensity,
-            highQuantile,
-            lockedMax !== undefined,
-          ),
-        }
-      }
-      return null
-    }
-
-    const startImageLoaderSettings = () => {
-      state.imageSettings = {
-        lockedIntensities: [undefined, undefined],
-        annotImageOpacity: 1.0,
-        opticalOpacity: 1.0,
-        colormap: 'Viridis',
-        scaleType: 'linear',
-        opacityMode: 'constant',
-        scaleBarColor: '#000000',
-        imagePosition: {
-          zoom: 1,
-          xOffset: 0,
-          yOffset: 0,
-        },
-        opticalSrc: null,
-        opticalTransform: null,
-        userScaling: [0, 1],
-        imageScaledScaling: [0, 1],
-        pixelAspectRatio: config.features.ignore_pixel_aspect_ratio ? 1
-          : getPixelSizeX() && getPixelSizeY() && getPixelSizeX() / getPixelSizeY() || 1,
-      }
-    }
-
     return () => {
       const isEmpty = state.x === undefined && state.y === undefined
 
@@ -759,81 +443,15 @@ export default defineComponent<DatasetBrowserProps>({
                   </div>
                 }
                 {
-                  state.imageSettings
-                  && <MainImageHeader
-                    class='dataset-comparison-grid-item-header dom-to-image-hidden'
-                    annotation={state.annotation}
-                    slot="title"
-                    hideTitle
-                    isActive={false}
-                    showOpticalImage={false}
-                    toggleOpticalImage={(e: any) => { }}
-                    resetViewport={startImageLoaderSettings}
-                    hasOpticalImage={false}
-                    colormap={state.imageSettings.colormap}
-                    onColormapChange={handleColormapChange}
-                    scaleType={state.imageSettings.scaleType}
-                    onScaleTypeChange={handleScaleTypeChange}
-                    onScaleBarColorChange={handleScaleBarColorChange}
-                  />
-                }
-                {
                   state.annotation
-                  && state.imageSettings
-                  && <div class='relative'>
-                    <MainImage
-                      keepPixelSelected
-                      annotation={state.annotation}
-                      opacity={1}
-                      hideColorBar
-                      imageLoaderSettings={state.imageSettings}
-                      imagePosition={state.imageSettings.imagePosition}
-                      applyImageMove={handleImageMove}
-                      colormap={state.imageSettings.colormap}
-                      scaleBarColor={state.imageSettings.scaleBarColor}
-                      scaleType={state.imageSettings.scaleType}
-                      userScaling={state.imageSettings.imageScaledScaling}
-                      pixelSizeX={getPixelSizeX()}
-                      pixelSizeY={getPixelSizeY()}
-                      {...{ on: { 'pixel-select': handlePixelSelect } }}
-                    />
-                    {
-                      state.imageSettings.intensity
-                      && <div class="ds-viewer-controls-wrapper  v-rhythm-3 sm-side-bar">
-                        <FadeTransition class="absolute top-0 right-0 mt-3 ml-3 dom-to-image-hidden">
-                          <div
-                            class="range-slider p-3 bg-gray-100 rounded-lg box-border shadow-xs">
-                            <RangeSlider
-                              class="ds-comparison-opacity-item"
-                              value={state.imageSettings.userScaling}
-                              min={0}
-                              max={1}
-                              step={0.01}
-                              style={state.rangeSliderStyle}
-                              onInput={handleUserScalingChange}
-                            />
-                            <div
-                              class="ds-intensities-wrapper">
-                              <IonIntensity
-                                intensities={state.imageSettings.intensity?.min}
-                                label="Minimum intensity"
-                                placeholder="min."
-                                onInput={(value: number) => handleIonIntensityChange(value, 'min')}
-                                onLock={(value: number) => handleIonIntensityLockChange(value, 'min')}
-                              />
-                              <IonIntensity
-                                intensities={state.imageSettings.intensity?.max}
-                                label="Maximum intensity"
-                                placeholder="man."
-                                onInput={(value: number) => handleIonIntensityChange(value, 'max')}
-                                onLock={(value: number) => handleIonIntensityLockChange(value, 'max')}
-                              />
-                            </div>
-                          </div>
-                        </FadeTransition>
-                      </div>
-                    }
-                  </div>
+                  && <SimpleIonImageViewer
+                    annotation={state.annotation}
+                    dataset={dataset.value}
+                    ionImageUrl={state.ionImageUrl}
+                    pixelSizeX={getPixelSizeX()}
+                    pixelSizeY={getPixelSizeY()}
+                    onPixelSelected={handlePixelSelect}
+                  />
                 }
               </div>
             </div>
